Move sass files into a sass folder
Following discussions within the team we feel it's more helpful to keep "source files" separate from "compiled files", so sass should be in its own folder.

We're using the name `sass` rather than `scss` because we feel that it's more standard within the community.

This allows us to remove and ignore the whole css directory.